Update datetime picker in Edit Course. Closes #181
I'm pretty sure that this is all the datetime pickers. Obviously, if one
comes up that I've missed, let me know.

Also, as a part of the process of fixing the datetime pickers, I've been
restyling forms to look more like native Bootstrap. At some point, I
hope that we can just use the `FormBuilderWithDateTimeInput` as our
default form initializer, as it contains a bunch of convenience methods
for adding help text and similar things.