Fix flash message for confirmation email sent
When devise sends an email with a validation/confirmation link after a
new user registers using the email/password "Sign up" option, it used to
say "You must be sign in or sign up to continue"

It showed this message because we were redirecting to "/", which
requires that the user is authorized. Now, it redirects to the sign in
page so that the devise flash message discussing what to do next doesn't
get clobbered by a chain of redirects.

Most code taken from the documentation at

https://github.com/plataformatec/devise/wiki/How-To:-Redirect-to-a-specific-page-on-successful-sign-up-%28registration%29

Fixes #478