Add correct timestamps to contents of zip file
I noticed that the contents of a zip file downloaded with the
"Download All Submissions" or "Download Final Submissions" links
would not preserve the timestamp of the submitted files. After
unzipping, the files would have the time stamp of when the zip
file was created, not the time when the file was handed in by
the student. This is useful information which should be preserved
when the zip archive is created.

Making this change has a few side effects which I also consider
to be improvements:

- previously a temporary file was created in /tmp with Tempfile.open
  which could end up hanging around after it is downloaded. What I
  do here is create a Zip::OutputStream and return that with send_data
  so no temporary file needs to be written to disk.

- also, the filename of the downloaded file used to be be something
  like: submissions20200220-79247-1k7kzv1.zip This is confusing to an
  instructor with multiple assessments or multiple courses, since
  the name doesn't specify which course or which assessment. Instead
  I use this filename: #{@course.name}_#{@course.semester}_#{@assessment.name}_submissions.zip