Send Tango unambiguous filenames for user handins (#1947)
* Export long handin filename

* Allow autogradeInputFiles to provide remote name

Allow autogradeInputFiles to provide a remoteFile attribute that will
be used for the filename on the tango server. If remoteFile is not
provided, the value of localFile will be used.

* Use full filename for tango handin files.

When uploading handins to tango, send the long filename (with the
student's email address) as the filename to use on the tango server

* sendJob remoteFile is always a basename

Unlike localFile, remoteFile can always be a basename, so remove
all the excess calls to File.basename